+++ src/main/java/com/example/examplemod/datagen/ModDataPackProvider.java
@@ -1,0 +1,22 @@
+package com.example.examplemod.datagen;
+
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+
+import com.example.examplemod.ExampleMod;
+
+import net.minecraft.core.HolderLookup.Provider;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.core.RegistrySetBuilder;
+import net.minecraft.data.PackOutput;
+import net.neoforged.neoforge.common.data.DatapackBuiltinEntriesProvider;
+
+public class ModDataPackProvider extends DatapackBuiltinEntriesProvider {
+    public static final RegistrySetBuilder BUILDER = new RegistrySetBuilder()
+            .add(Registries.PAINTING_VARIANT, ModPaintings::bootstrap);
+
+    public ModDataPackProvider(PackOutput output, CompletableFuture<Provider> registries) {
+        super(output, registries, BUILDER, Set.of(ExampleMod.MOD_ID));
+    }
+
+}